Make events/async macro better.
Events now don't rely on channels, but rather a loop within a Future.
This is because the channel.receive was blocking when I would rather it
didn't.

`async` macro now generates the async method with the original method
name, and renames the sync method to `{{ method_name }}_sync`. This way
that method is still accessible if desired. Better API imo.